refactor: replace preferences store with domain config store in components
