+++ apps/web/src/services/advanced/components/PreferencesContainer.vue
@@ -11,12 +11,12 @@ import { i18n } from '@/translations';
 
 import { makeAdminRouteName } from '@/router/helpers/route-helper';
 
-import { usePreferencesStore } from '@/store/preferences/preferences-store';
+import { useDomainConfigStore } from '@/store/config/domain-config-store';
 
 import { ADVANCED_ROUTE } from '@/services/advanced/routes/route-constant';
 
 
-const domainConfigStore = usePreferencesStore();
+const domainConfigStore = useDomainConfigStore();
 
 const route = useRoute();
