+++ apps/web/src/services/advanced/pages/admin/AdminDomainSettingsAppearancePage.vue
@@ -8,13 +8,13 @@ import {
 
 import { i18n } from '@/translations';
 
-import { usePreferencesStore } from '@/store/preferences/preferences-store';
+import { useDomainConfigStore } from '@/store/config/domain-config-store';
 
 import ErrorHandler from '@/common/composables/error/errorHandler';
 import { usePageEditableStatus } from '@/common/composables/page-editable-status';
 
 const router = useRouter();
-const domainConfigStore = usePreferencesStore();
+const domainConfigStore = useDomainConfigStore();
 const domainConfigGetters = domainConfigStore.getters;
 
 const { hasReadWriteAccess } = usePageEditableStatus();
